==> UTIL/02_Finding_conserved_genes.rmd <==
---
title: "02_Finding_conserved_genes.rmd"
author: "Soren Maret"
date: "03.29.25"
output: github_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(dplyr)
library(tidyr)
```
GOAL: Identify genes that changed in humans and mice stem cells in response to dox.To do this we want to compare our lists of sig genes from mouse and humans to a list of orthologues. This is in effect the same thing as taking the union between the 2 lists. 
```{r loading in data}
# loading realivent dataframes:
# Loading orthologues list
load ("/scratch/Shares/rinnclass/MASTER_CLASS/STUDENTS/genehomies/RESULTS/MOUSEVSHUMAN/orthologs.RData")
# Loading human sig genes
load ("/scratch/Shares/rinnclass/MASTER_CLASS/STUDENTS/genehomies/RESULTS/HUMAN/filtered_res_df_2")
#loading mouse sig genes
load ("/scratch/Shares/rinnclass/MASTER_CLASS/STUDENTS/soma3841/MASTER_CLASS/lessons/06_Differential_expression_analyses/results/DESEQ_results.rdata")
```
First things first we need to make our data compatable. right now if we try and filter our dataframe we wont get any genes back because of the decimal values that are present in our sig gene dataframes will block our downstream analysis. 
```{r making dataframes for analysis}
# lets make a new dataframe with gene_id and gene_name
mouse_gene_df <- filtered_res_df[, c("gene_id", "gene_name")]

head(mouse_gene_df)

#we now need to remove the decimal values from our gene_id collumn for orthology analysis
mouse_gene_df <- mouse_gene_df %>%
  mutate(gene_id = sub("\\..*", "", gene_id))
# View result
head(mouse_gene_df)
#great!! that worked. now lets do the same for our Human d
```

```{r making dataframes for analysis human}
human_gene_df <- filtered_res_df_2[, c("gene_id", "gene_name")]
#lets check
head(human_gene_df)

#we now need to remove the decimal values from our gene_id collumn for orthology analysis
human_gene_df <- human_gene_df %>%
  mutate(gene_id = sub("\\..*", "", gene_id))
# view result
head(human_gene_df)
```

```{r filtering our orthologs}
# Filter the orthologue table
conserved_changing_orthologues <- orthologs_filtered %>%
  filter(Gene.stable.ID %in% human_gene_df$gene_id,
         Mouse.gene.stable.ID %in% mouse_gene_df$gene_id)

# View the result
head(conserved_changing_orthologues)

#note how our dataframe is small than either of our 2 input dataframes suggesting that our analysis worked. 
```
#RESULT: 

!! We have extracted 276 genes that both signifigantly changed in our human and mouse samples !!

!! We have several duplicated genes names !!
```{r depreplicating our data}
#lets go and get rid of our duplicates based on the gene name
# Remove duplicate rows based on "Gene.name"
conserved_changing_orthologues_dereplicated <- conserved_changing_orthologues %>%
  distinct(Gene.name, .keep_all = TRUE)
# WHOA that was a nearly 10x reduction. lets see what these genes are!
cat(conserved_changing_orthologues_dereplicated$Gene.name, sep = "\n")
# copy and paste into https://maayanlab.cloud/Enrichr/ 
```
Lets save our files!
```{r saving our results}
save(human_gene_df, mouse_gene_df, conserved_changing_orthologues, conserved_changing_orthologues_dereplicated, file = "/scratch/Shares/rinnclass/MASTER_CLASS/STUDENTS/genehomies/RESULTS/MOUSEVSHUMAN/orthologs_results.RData")
#checking... 
load("/scratch/Shares/rinnclass/MASTER_CLASS/STUDENTS/genehomies/RESULTS/MOUSEVSHUMAN/orthologs_results.RData")
```
